% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acker-package.R
\docType{package}
\name{acker-package}
\alias{acker}
\alias{acker-package}
\title{acker: A Shiny App for Crop Simulation}
\description{
Shinyapp to simulate corn growth in an educational context.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/assignUser/acker/issues}
  \item Report bugs at \url{https://github.com/assignUser/acker/issues/issues}
}

}
\author{
\strong{Maintainer}: Jacob Wujciak-Jens \email{jacob@wujciak} (\href{https://orcid.org/0000-0002-7281-3989}{ORCID})

}
\keyword{internal}
