% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drought_index}
\alias{drought_index}
\title{DWD Climate Data Center (CDC): Annual grids of drought index (de Martonne) over Germany, version v1.0}
\format{
An object of class \code{SpatRaster} of dimension 712 x 917 x 27.
}
\usage{
drought_index
}
\description{
DWD Climate Data Center (CDC): Annual grids of drought index (de Martonne) over Germany, version v1.0
}
\keyword{datasets}
