% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{min_data}
\alias{min_data}
\title{TODO document data sets / make it load on attach}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 365 rows and 9 columns.
}
\usage{
min_data
}
\description{
data from years with min max and median DI in germany
}
\details{
year drought_index
2002      51.79878
2005      41.32579
2018      28.30215
}
\keyword{datasets}
