% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{min_di_weather}
\alias{min_di_weather}
\title{Data from years with min max and median DI in Fermany}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 365 rows and 9 columns.
}
\usage{
min_di_weather
}
\description{
year drought_index
2002      51.79878
2005      41.32579
2018      28.30215
}
\seealso{
[acker::drought_index]
}
\keyword{datasets}
