% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acker-package.R
\docType{package}
\name{acker-package}
\alias{acker}
\alias{acker-package}
\title{acker: A Shiny App for Crop Simulation}
\description{
TODO
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/assignUser/acker/issues}
  \item Report bugs at \url{https://github.com/assignUser/acker/issues/issues}
}

}
\author{
\strong{Maintainer}: Jacob Wujciak-Jens \email{jacob@wujciak}

}
\keyword{internal}
